+++ model_utils/models.py
@@ -67,6 +67,20 @@ class StatusModel(models.Model):
     status = StatusField(_('status'))
     status_changed = MonitorField(_('status changed'), monitor='status')
 
+    def save(self, *args, **kwargs):
+        """
+        Overriding the save method in order to make sure that
+        status_changed field is updated even if it is not given as
+        a parameter to the update field argument.
+        """
+        if (
+            'update_fields' in kwargs
+            and 'status' in kwargs['update_fields']
+            and 'status_changed' not in kwargs['update_fields']
+        ):
+            kwargs['update_fields'] += ['status_changed']
+        super().save(*args, **kwargs)
+
     class Meta:
         abstract = True